==> 4_Dosage consequences.R <==
setwd("~/Library/CloudStorage/OneDrive-UniversityofEdinburgh/Quantitative_Resistance/Quantitative_Resistance")
library(foreach)
library(doParallel)
library(deSolve)
library(dplyr)
library(tidyverse)

#######################################################################################################################
# MODEL

mutation_rate<-10^-8 # Minimum mutation rate per division
V_max<-0.5 # Maximum growth rate per hour
cmax<-0.025 # Maximum cost of resistance
kp<-5*10^-4 # Immune cell killing rate per hour
eta<-0.003 # Immune cell recruitment rate (changed to make the infection last 100 days)
ka_max<-0.03125 # Maximum antibiotic killing rate per hour (hourly death rate generated by the antibiotic = -0.75)
d<-0.1 # Antibiotic decay rate
time_to_antibiotic<-24*3 # Once bacteria hit maximum (~ 3 days)

b_max<-10^10 # Maximum bacterial cell density
mmax<-10 # Maximum MIC
step_size<-1 # For euler integration method
bacteria_start_val<-1
minimum_bacteria<-1 
immune_start_val<-1
runtime<-110*24
one_day<-24
dose_time<-24
course_length<-110

model<-function(times,state,pars){
  with(as.list(c(state,pars)),{
    
    if (n==1) {
      N<-b1
      db1 <- (bacterial_parameters$psi[1]*(1-N/b_max) - P*kp - A*bacterial_parameters$ka[1]) * b1
      if(b1<minimum_bacteria){db1<- -b1*(1/step_size)}
      bacteria_strains<-db1
      
      
    } else if (n==2) {
      N<-b1+b2
      if (b1>=minimum_bacteria){mutations<-rpois(1,lambda=bacterial_parameters$psi[1]*(1-N/b_max)*bacterial_parameters$mew*b1)} else {mutations<-0}
      db1 <- (bacterial_parameters$psi[1]*(1-N/b_max) - P*kp - A*bacterial_parameters$ka[1]) * b1 - mutations
      db2 <- (bacterial_parameters$psi[2]*(1-N/b_max) - P*kp - A*bacterial_parameters$ka[2]) * b2 + mutations
      if(b1<minimum_bacteria){db1<- -b1*(1/step_size)}
      if(b2<minimum_bacteria & b2!=0){db2<- -b2*(1/step_size)}
      if(b2<=5 & sample(1:(1/(0.5*step_size)), 1, replace=TRUE)==1){db2<- -b2*(1/step_size)}
      bacteria_strains<-c(db1,db2)
      
    } else {
      N <- 0
      for(i in 1:n){
        current <- eval(as.symbol(paste0("b",i)))
        N <- N + current}
      if (b1>=minimum_bacteria){mutations_out<-rpois(1,lambda=bacterial_parameters$psi[1]*(1-N/b_max)*bacterial_parameters$mew*b1)} else {mutations_out<-0}
      db1 <- (bacterial_parameters$psi[1]*(1-N/b_max) - P*kp - A*bacterial_parameters$ka[1]) * b1 - mutations_out
      if(b1<minimum_bacteria){db1<- -b1*(1/step_size)}
      for(i in 2:(n-1)) {
        mutations_in<-mutations_out
        bi<-eval(as.symbol(paste0("b",i)))
        if(bi>=minimum_bacteria){mutations_out<-rpois(1,lambda=bacterial_parameters$psi[i]*(1-N/b_max)*bacterial_parameters$mew*bi)} else{mutations_out<-0}
        assign(paste0("db",i), (bacterial_parameters$psi[i]*(1-N/b_max) - P*kp - A*bacterial_parameters$ka[i]) * bi - mutations_out + mutations_in)
        if(bi<minimum_bacteria & bi!=0){assign(paste0("db",i),-bi*(1/step_size))}
        if(bi<=5 & sample(1:(1/(0.5*step_size)), 1, replace=TRUE)==1){assign(paste0("db",i),-bi*(1/step_size))}
      }
      mutations_in<-mutations_out
      bn<-eval(as.symbol(paste0("b",n)))
      assign(paste0("db",n), (bacterial_parameters$psi[n]*(1-N/b_max) - P*kp - A*bacterial_parameters$ka[n]) * bn + mutations_in)
      if(bn<minimum_bacteria & bn!=0){assign(paste0("db",n),-bn*(1/step_size))}
      if(bn<=5 & sample(1:(1/(0.5*step_size)), 1, replace=TRUE)==1){assign(paste0("db",n),-bn*(1/step_size))}
      bacteria_strains<-vector()
      for(i in 1:n){bacteria_strains[i]<-eval(as.symbol(paste0("db",i)))}
      
    }
    
    dA <- -d*A # Rate of change of antibiotic
    dP<-eta*P*(N/(N+10^3)) # Rate of change of immune cells (bacteria density dependent)
    list(c(bacteria_strains,dA,dP))
  })
}

#######################################################################################################################
# EXAMPLE DYNANMICS

df_n_dose_runs<-data.frame(n=c(2,2,10,10),dose=c(0,40,0,40))
examples_data<-data.frame()
for(j in 1:nrow(df_n_dose_runs)){
    
  n<-df_n_dose_runs$n[j]
  dose<-df_n_dose_runs$dose[j]

  bacterial_parameters<-data.frame()
  for (k in 1:n){
    mic<-(mmax^(1/(n-1)))^(k-1) # MIC for each strain
    cost<-(cmax/(mmax-1))*(mic-1)  # Cost of resistance for each strain
    psi<-V_max*(1-cost) # Growth rate for each strain
    ka<-ka_max/mic # Antibiotic killing rate for each strain
    mew<-mutation_rate*(n-k) # Mutation rate for each strain
    bacterial_parameters<-rbind(bacterial_parameters,data.frame(genotype=k,cost,psi,mic,ka,mew))
  }
  
  antibiotic_taken<-seq(time_to_antibiotic,time_to_antibiotic+(one_day*course_length),dose_time)
  eventdat <- data.frame(var = "A", time = antibiotic_taken, value = dose, method = "add")
  times<-seq(0,runtime,length.out=runtime+1)
  pars<-c(bacterial_parameters, kp, d, eta, b_max)
  if(n==1){state<-c(b1=bacteria_start_val,A=0,P=immune_start_val)}else{state<-c(b=c(bacteria_start_val,rep(0,n-1)),A=0,P=immune_start_val)}
  out<-as.data.frame(ode(y=state,times=times,func=model,parms=pars,method='euler',hini=step_size,events = list(data = eventdat)))
  out[out<1]<-0

  df_plot<-gather(out,"genotype","frequency",2:(n+1))
  df_plot$genotype<-as.numeric(gsub("b","",df_plot$genotype))
  df_plot$genotype <- as.character(df_plot$genotype)
  bacterial_parameters$genotype <- as.character(bacterial_parameters$genotype)
  df_plot<-full_join(df_plot,dplyr::select(bacterial_parameters,genotype,mic),by="genotype")
  df_plot$time<-df_plot$time/24
  df_plot<-df_plot %>% mutate(facet=paste0("Strains: ", n, "  Dose: ",dose))
  examples_data<-rbind(examples_data,df_plot)
  }
write.csv(examples_data,file="within_host_results_examples.csv",row.names = FALSE)

#######################################################################################################################
# ALL N AND DOSE

n_list<-2:10
dose_list<-0:40
runs<-1:10
n_list_rep<-data.frame(n=rep(n_list,times=length(runs)*length(dose_list))) %>% arrange(n)
dose_list_rep<-bind_rows(replicate(length(n_list), data.frame(dose=rep(dose_list,times=length(runs))) %>% arrange(dose), simplify = FALSE))
runs_rep<-bind_rows(replicate(length(n_list)*length(dose_list), data.frame(run=runs), simplify = FALSE))
df_n_dose_runs<-cbind(n_list_rep,dose_list_rep,runs_rep)
  
print(Sys.time())
n.cores <- parallel::detectCores() - 1
my.cluster <- parallel::makeCluster(n.cores,type = "FORK")
doParallel::registerDoParallel(cl = my.cluster)
all_results<-foreach (j = 1:nrow(df_n_dose_runs),.combine=rbind) %dopar% {
  
  n<-df_n_dose_runs$n[j]
  dose<-df_n_dose_runs$dose[j]
  run<-df_n_dose_runs$run[j]
  
  bacterial_parameters<-data.frame()
  for (k in 1:n){
    mic<-(mmax^(1/(n-1)))^(k-1) # MIC for each strain
    cost<-(cmax/(mmax-1))*(mic-1) # Cost of resistance for each strain
    psi<-V_max*(1-cost) # Growth rate for each strain
    ka<-ka_max/mic # Antibiotic killing rate for each strain
    mew<-mutation_rate*(n-k) # Mutation rate for each strain
    bacterial_parameters<-rbind(bacterial_parameters,data.frame(genotype=k,cost,psi,mic,ka,mew))
  }
  
  antibiotic_taken<-seq(time_to_antibiotic,time_to_antibiotic+(one_day*course_length),dose_time)
  eventdat <- data.frame(var = "A", time = antibiotic_taken, value = dose, method = "add")
  times<-seq(0,runtime,length.out=runtime+1)
  pars<-c(bacterial_parameters, kp, d, eta, b_max)
  if(n==1){state<-c(b1=bacteria_start_val,A=0,P=immune_start_val)}else{state<-c(b=c(bacteria_start_val,rep(0,n-1)),A=0,P=immune_start_val)}
  out<-as.data.frame(ode(y=state,times=times,func=model,parms=pars,method='euler',hini=step_size,events = list(data = eventdat)))
  out[out<1]<-0
  
  all_results<-filter(out,time>=time_to_antibiotic+1)[2:(n+1)] %>% summarise_all(funs(sum)) %>% gather(strain,freq) %>% cbind(data.frame(mic=bacterial_parameters$mic,n=n,dose=dose,run=run))
}

parallel::stopCluster(cl = my.cluster)
print(Sys.time())
write.csv(all_results,file="within_host_results.csv",row.names = FALSE)
